Added files via upload
The ML Class Library (mllib for short) is a python class library for training and deploying machine learning algorithms. The library provides a higher level of abstraction over ML libs such as sci-kit learn and pandas. The project consists of 5 major classes that are meant to reflect the stages in an ML project
- Projects (organize data prep and training as a project. 
- exploreData: tools for exploring, plotting and evaluating data
- cleanData: Tools to clean and feature engineer the data
- prepData: Prepare the data for ML training
- trainModels: Run training and fits, also run predicts on saved models